requirements.txt
pandas
psycopg2-binary
yfinance
cloudscraper
lxml
fastapi
uvicorn[standard]